Added python example on how to connect and communicate to the main app. Can run on Raspberry Pi.
